enable sq tuning for both quant_level is auto or 1
